Use world context instead of trade context for eth price and osqth bal (#67)
