+++ packages/frontend/src/components/Charts/LongChart.tsx
@@ -13,11 +13,10 @@ import InfoIcon from '@material-ui/icons/InfoOutlined'
 import dynamic from 'next/dynamic'
 import Image from 'next/image'
 import React, { useEffect, useMemo, useState } from 'react'
-import { useTrade } from '@context/trade'
+import { useWorldContext } from '@context/world'
 import ComparisonChart from '../../../public/images/ComparisonChart.svg'
 import { graphOptions } from '../../constants/diagram'
 import { Links, Tooltips } from '../../constants/enums'
-import { useWorldContext } from '../../context/world'
 import IV from '../IV'
 import { SqueethTab, SqueethTabs } from '../Tabs'
 import LongSqueethPayoff from './LongSqueethPayoff'
@@ -97,7 +96,7 @@ export function LongChart() {
   const [tradeType, setTradeType] = useState(0)
 
   const classes = useStyles()
-  const { ethPrice } = useTrade()
+  const { ethPrice } = useWorldContext()
 
   useEffect(() => {
     if (tradeType === 0) setMode(ChartType.PNL)
