+++ packages/frontend/src/components/Lp/LPTable.tsx
@@ -21,7 +21,7 @@ import { useEffect, useState } from 'react'
 import { SecondaryTab, SecondaryTabs } from '../../components/Tabs'
 import { Tooltips, UniswapIFrameOpen, UniswapIFrameClose } from '@constants/enums'
 import { useSqueethPool } from '@hooks/contracts/useSqueethPool'
-import { useTrade } from '@context/trade'
+import { useWorldContext } from '@context/world'
 import { useLPPositions } from '@hooks/usePositions'
 import { inRange } from '@utils/calculations'
 import { UniswapIframe } from '../Modal/UniswapIframe'
@@ -117,7 +117,7 @@ export const LPTable: React.FC<LPTableProps> = ({ isLPage, pool }) => {
   const classes = useStyles()
   const { activePositions, closedPositions, loading: lpLoading } = useLPPositions()
   const [activeTab, setActiveTab] = useState(0)
-  const { ethPrice } = useTrade()
+  const { ethPrice } = useWorldContext()
   const { getWSqueethPositionValue, isWethToken0 } = useSqueethPool()
   const { networkId } = useWallet()
