+++ packages/frontend/src/components/PositionCard.tsx
@@ -10,6 +10,7 @@ import React, { useCallback, useEffect, useMemo, useState } from 'react'
 import { usePnL, usePositions } from '@hooks/usePositions'
 import { Tooltips } from '@constants/enums'
 import { useTrade } from '@context/trade'
+import { useWorldContext } from '@context/world'
 import { PositionType, TradeType } from '../types'
 import { useController } from '@hooks/contracts/useController'
 import { toTokenAmount } from '@utils/calculations'
@@ -171,6 +172,7 @@ const PositionCard: React.FC<PositionCardType> = ({ tradeCompleted }) => {
     existingCollat,
     loading: isPositionLoading,
     isLP,
+    isLong,
   } = usePositions()
   const { liquidations } = useVaultLiquidations(Number(vaultId))
   const {
@@ -181,8 +183,8 @@ const PositionCard: React.FC<PositionCardType> = ({ tradeCompleted }) => {
     tradeSuccess,
     setTradeSuccess,
     tradeType,
-    ethPrice,
   } = useTrade()
+  const { ethPrice } = useWorldContext()
   const tradeAmount = new BigNumber(tradeAmountInput)
   const { index } = useController()
   const [fetchingNew, setFetchingNew] = useState(false)
@@ -348,7 +350,9 @@ const PositionCard: React.FC<PositionCardType> = ({ tradeCompleted }) => {
                     <Typography variant="caption" color="textSecondary" style={{ fontWeight: 500 }}>
                       Unrealized P&L
                     </Typography>
-                    <Tooltip title={Tooltips.UnrealizedPnL}>
+                    <Tooltip
+                      title={isLong ? Tooltips.UnrealizedPnL : `${Tooltips.UnrealizedPnL}. ${Tooltips.ShortCollateral}`}
+                    >
                       <InfoIcon fontSize="small" className={classes.infoIcon} />
                     </Tooltip>
                   </div>
@@ -380,7 +384,9 @@ const PositionCard: React.FC<PositionCardType> = ({ tradeCompleted }) => {
                   <Typography variant="caption" color="textSecondary" style={{ fontWeight: 500 }}>
                     Realized P&L
                   </Typography>
-                  <Tooltip title={Tooltips.RealizedPnL}>
+                  <Tooltip
+                    title={isLong ? Tooltips.RealizedPnL : `${Tooltips.RealizedPnL}. ${Tooltips.ShortCollateral}`}
+                  >
                     <InfoIcon fontSize="small" className={classes.infoIcon} />
                   </Tooltip>
                 </div>
