+++ packages/frontend/src/components/Trade/History.tsx
@@ -5,7 +5,7 @@ import { EtherscanPrefix } from '../../constants'
 import { TransactionType } from '@constants/enums'
 import { useWallet } from '@context/wallet'
 import { useController } from '@hooks/contracts/useController'
-import { useTrade } from '@context/trade'
+import { useWorldContext } from '@context/world'
 import { useTransactionHistory } from '@hooks/useTransactionHistory'
 import { useUsdAmount } from '@hooks/useUsdAmount'
 const useStyles = makeStyles((theme) =>
@@ -43,7 +43,7 @@ const useStyles = makeStyles((theme) =>
 const History: React.FC = () => {
   const { transactions } = useTransactionHistory()
   const { networkId } = useWallet()
-  const { ethPrice } = useTrade()
+  const { ethPrice } = useWorldContext()
   const { normFactor: normalizationFactor } = useController()
   const classes = useStyles()
   const { getUsdAmt } = useUsdAmount()
